API-4199: moved Acceptance Tests into root
