+++ project/plugins.sbt
@@ -3,10 +3,10 @@ resolvers += "HMRC Releases" at "https://dl.bintray.com/hmrc/releases"
 
 resolvers += "Typesafe Releases" at "http://repo.typesafe.com/typesafe/releases/"
 
-addSbtPlugin("uk.gov.hmrc" % "sbt-auto-build" % "2.6.0")
-addSbtPlugin("uk.gov.hmrc" % "sbt-git-versioning" % "2.1.0")
-addSbtPlugin("uk.gov.hmrc" % "sbt-distributables" % "2.0.0")
+addSbtPlugin("uk.gov.hmrc" %% "sbt-auto-build" % "2.6.0")
+addSbtPlugin("uk.gov.hmrc" %% "sbt-git-versioning" % "2.1.0")
+addSbtPlugin("uk.gov.hmrc" %% "sbt-distributables" % "2.0.0")
 addSbtPlugin("org.scoverage" %% "sbt-scoverage" % "1.5.1")
 addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "1.0.0")
-addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.6.24")
-addSbtPlugin("uk.gov.hmrc" % "sbt-artifactory" % "1.0.0")
+addSbtPlugin("com.typesafe.play" %% "sbt-plugin" % "2.6.24")
+addSbtPlugin("uk.gov.hmrc" %% "sbt-artifactory" % "1.0.0")
